Fix the dead buttons: a press was gated on a grab Advance had already retired
Every button in the tool-options widget and the context menu was dead. Mode
segments, toggles, swatches, Apply, Cancel -- none of them responded. The
slider worked, which is what made this look like a drawing or a layout fault
rather than an input one.

`Pressed` asked `Holding(Target)` on the tick the contact was released. But
`ControlIndex::Advance` runs at the top of the frame, before any control has
recorded, and when the contact is no longer held it retires the grab into
`ReleasedControl`. So by the time a control asked whether it was still being
held, the answer was always no. The press could not fire; there was no timing
window in which it could have.

The slider escaped because it grabs on `ContactPressed` and drives on `Holding`
during the drag. It never consults the release, so the one path that was broken
was the one path it does not use.

`Pressed` now reads the release the index actually recorded:

    Hovered && Interaction->Released(Target)
            && Interaction->ReleasedControlPart(Target) == ControlPart::Body

The hover test stays, so dragging off a control and releasing does not press it.

Three copies of this function existed. `ToolOptionsWidget::Pressed` already read
`Released` and was correct -- which is why the collapse and close chevrons worked
while nothing inside the widget did. `OptionControls` and `ToolContextMenu` both
read `Holding`. Two copies of one function disagreed, every unit proof passed,
and no claim compared them. `ToolOptionsWidgetProof` now asserts all three read
the release, against source text, and links the real `ControlIndex.cpp` so the
press claims exercise the actual grab-and-release rotation. A proof that
reimplements the rule it tests would have reproduced this defect faithfully and
passed.

The dimension filter gates now, it does not hide
------------------------------------------------
The construction operations were unreachable for a second reason. `Presented`
refuses a hidden tool outright and `ShowGated` cannot bypass that, so a
dimension filter answering `Hidden` removed all five Sketch Modify tools
whenever nothing was selected. `VisibleCount` then fell to zero and the rail
skipped the entire band. The artist could not see the operations or the band
that holds them -- and reaching the dimension that would reveal them required
selecting an edge, which is the thing they had gone to the catalogue to find an
operation for.

The dimension check moved into `Gated`, so an inapplicable tool is dimmed and
still visible. `Hidden` keeps only `Raising` against a solid, which is "never"
rather than "not yet" and so is correctly removed instead of dimmed forever.

The sandbox was rejecting 35 units for reasons unrelated to any change
----------------------------------------------------------------------
Every `ExternalPackages/` submodule was an empty directory here and VULKAN_SDK
pointed at the parse stub, so 35 of 255 translation units failed on every run.
Those 35 were the whole SlateVulkan device layer and most of SlateUI: a change
could break the interface layer outright and the script would report the same
35 failures it always reported. A standing baseline of expected failures is not
a baseline, it is a blind spot.

Submodules populated at their recorded commits and the real Vulkan headers in
place, the build is now 255 of 255 accepted, and the setup is written down at
the top of the script so the blind spot does not come back.

Also cleared the warnings from the last Windows build -- a shadowed `Root` in
FontLoader, a silent double-to-Real32 narrowing in EditorHost, and eight
uint64-to-uint32 conversions in CodexSceneProxy -- and stopped the widget proof
leaving object files beside its source, where five of them had been committed.

+++ Engine/SlateWorkspace/Discipline/CodexSceneProxy/Source/CodexSceneProxy.cpp
@@ -236,7 +236,7 @@ bool ResolveSelectedSceneMeshPivot(const WorkspaceCodex& Scene,
     const StableRowIdentity Identity = Storage.Rows[Applied.EntityTaken].Identity;
     if (Identity < 6200u)
         return false;
-    const std::uint32_t SceneIndex = Identity - 6200u;
+    const std::uint32_t SceneIndex = static_cast<std::uint32_t>(Identity - 6200u);
     if (SceneIndex >= Scene.Scene.size() || Scene.Scene[SceneIndex].Subject != CodexSceneSubject::Geometry)
         return false;
     Pivot = CodexScenePosition(Scene.Scene[SceneIndex], UnitScale);
@@ -262,7 +262,7 @@ bool SelectSceneMeshAtPointer(const PlaneExtent& Extent,
         const StableRowIdentity Identity = Storage.Rows[Row].Identity;
         if (Identity < 6200u)
             continue;
-        const std::uint32_t SceneIndex = Identity - 6200u;
+        const std::uint32_t SceneIndex = static_cast<std::uint32_t>(Identity - 6200u);
         if (SceneIndex >= Scene.Scene.size())
             continue;
         const CodexSceneEntry& Entry = Scene.Scene[SceneIndex];
@@ -479,7 +479,7 @@ void SeedSceneDirectoryTransformsFromCodex(const WorkspaceCodex& Scene,
         const StableRowIdentity Identity = Storage.Rows[Row].Identity;
         if (Identity < 6200u)
             continue;
-        const std::uint32_t SceneIndex = Identity - 6200u;
+        const std::uint32_t SceneIndex = static_cast<std::uint32_t>(Identity - 6200u);
         if (SceneIndex >= Scene.Scene.size())
             continue;
         const CodexSceneEntry& Entry = Scene.Scene[SceneIndex];
@@ -505,7 +505,7 @@ void SynchroniseCodexTransformsFromSceneDirectory(WorkspaceCodex& Scene,
         const StableRowIdentity Identity = Storage.Rows[Row].Identity;
         if (Identity < 6200u)
             continue;
-        const std::uint32_t SceneIndex = Identity - 6200u;
+        const std::uint32_t SceneIndex = static_cast<std::uint32_t>(Identity - 6200u);
         if (SceneIndex >= Scene.Scene.size())
             continue;
         CodexSceneEntry& Entry = Scene.Scene[SceneIndex];
